+++ iftttie/channels/buienradar.py
@@ -9,7 +9,7 @@ from loguru import logger
 
 from iftttie.channels.base import BaseChannel
 from iftttie.context import Context
-from iftttie.types import Event, Unit
+from iftttie.types_ import Event, Unit
 
 url = 'https://api.buienradar.nl/data/public/2.0/jsonfeed'
 headers = [('Cache-Control', 'no-cache')]